CORE-13661 Chainable tests refactored
