+++ e2e-tests/cpbs/ledger-utxo-advanced-chainable-demo-app/src/main/kotlin/com/r3/corda/demo/utxo/chainable/workflow/testing/ChainableContractTestFlow.kt
@@ -51,166 +51,4 @@ class ChainableContractTestFlow : ClientStartableFlow {
     }
 }
 
-class ChainableContractCreateTestFlow(private val rule: String) : SubFlow<List<StateAndRef<*>>> {
-
-    @CordaInject
-    private lateinit var digestService: DigestService
-
-    @CordaInject
-    private lateinit var memberLookup: MemberLookup
-
-    @CordaInject
-    private lateinit var notaryLookup: NotaryLookup
-
-    @CordaInject
-    private lateinit var utxoLedgerService: UtxoLedgerService
-
-    @Suspendable
-    override fun call(): List<StateAndRef<*>> {
-        val key = memberLookup.myInfo().firstLedgerKey
-        val outputs = when (rule) {
-            "CONTRACT_RULE_CREATE_OUTPUTS" -> listOf(MyContractState(UUID.randomUUID()))
-            "CONTRACT_RULE_CREATE_POINTERS" -> {
-                listOf(
-                    MyChainableState(
-                        UUID.randomUUID(),
-                        key,
-                        StaticPointer(
-                            StateRef(digestService.hash(byteArrayOf(1, 2, 3, 4), DigestAlgorithmName.SHA2_256), 0),
-                            MyChainableState::class.java
-                        )
-                    ),
-                    MyChainableState(UUID.randomUUID(), key, null)
-                )
-            }
-            "VALID" -> {
-                listOf(
-                    MyChainableState(UUID.randomUUID(), memberLookup.myInfo().firstLedgerKey, null),
-                    MyChainableState(UUID.randomUUID(), memberLookup.myInfo().firstLedgerKey, null),
-                    MyContractState(UUID.randomUUID())
-                )
-            }
-            else -> throw IllegalArgumentException("Invalid rule type passed in")
-        }
-        val transaction = utxoLedgerService.createTransactionBuilder()
-            .setNotary(notaryLookup.notaryServices.first().name)
-            .addOutputStates(outputs)
-            .addSignatories(key)
-            .addCommand(MyChainableContract.Create())
-            .setTimeWindowUntil(Instant.now().plus(10, ChronoUnit.DAYS))
-            .toSignedTransaction()
-
-        return utxoLedgerService.finalize(transaction, emptyList()).transaction.outputStateAndRefs
-    }
-}
-
-class ChainableContractUpdateTestFlow(
-    private val rule: String,
-    private val stateAndRefs: List<StateAndRef<*>>
-) : SubFlow<List<StateAndRef<*>>> {
-
-    @CordaInject
-    private lateinit var memberLookup: MemberLookup
-
-    @CordaInject
-    private lateinit var notaryLookup: NotaryLookup
-
-    @CordaInject
-    private lateinit var utxoLedgerService: UtxoLedgerService
-
-    @Suspendable
-    override fun call(): List<StateAndRef<*>> {
-        val key = memberLookup.myInfo().firstLedgerKey
-        val (inputs, outputs) = when (rule) {
-            "CONTRACT_RULE_UPDATE_INPUTS" -> {
-                emptyList<StateRef>() to listOf(MyContractState(UUID.randomUUID()))
-            }
-            "CONTRACT_RULE_UPDATE_OUTPUTS" -> {
-                stateAndRefs.map { it.ref } to emptyList()
-            }
-            "CONTRACT_RULE_UPDATE_POINTERS" -> {
-                stateAndRefs.map { it.ref } to stateAndRefs
-                    .filter { it.state.contractState is MyChainableState }
-                    .map { (it.state.contractState as MyChainableState).copy(previousStatePointer = null) }
-            }
-            "CONTRACT_RULE_UPDATE_EXCLUSIVE_POINTERS" -> {
-                val stateRef = stateAndRefs.first().ref
-                stateAndRefs.map { it.ref } to stateAndRefs
-                    .filter { it.state.contractState is MyChainableState }
-                    .map {
-                        (it.state.contractState as MyChainableState).copy(
-                            previousStatePointer = StaticPointer(
-                                stateRef,
-                                MyChainableState::class.java
-                            )
-                        )
-                    }
-            }
-            "VALID" -> {
-                stateAndRefs.map { it.ref } to stateAndRefs
-                    .filter { it.state.contractState is MyChainableState }
-                    .map {
-                        (it.state.contractState as MyChainableState).copy(
-                            previousStatePointer = StaticPointer(
-                                it.ref,
-                                MyChainableState::class.java
-                            )
-                        )
-                    }
-            }
-            else -> throw IllegalArgumentException("Invalid rule type passed in")
-        }
-        val transaction = utxoLedgerService.createTransactionBuilder()
-            .setNotary(notaryLookup.notaryServices.first().name)
-            .addInputStates(inputs)
-            .addOutputStates(outputs)
-            .addSignatories(key)
-            .addCommand(MyChainableContract.Update())
-            .setTimeWindowUntil(Instant.now().plus(10, ChronoUnit.DAYS))
-            .toSignedTransaction()
-
-        return utxoLedgerService.finalize(transaction, emptyList()).transaction.outputStateAndRefs
-    }
-}
-
-class ChainableContractDeleteTestFlow(
-    private val rule: String,
-    private val stateAndRefs: List<StateAndRef<*>>
-) : SubFlow<List<StateAndRef<*>>> {
-
-    @CordaInject
-    private lateinit var memberLookup: MemberLookup
-
-    @CordaInject
-    private lateinit var notaryLookup: NotaryLookup
-
-    @CordaInject
-    private lateinit var utxoLedgerService: UtxoLedgerService
-
-    @Suspendable
-    override fun call(): List<StateAndRef<*>> {
-        val key = memberLookup.myInfo().firstLedgerKey
-        val inputs = when (rule) {
-            "CONTRACT_RULE_DELETE_INPUTS" -> {
-                stateAndRefs
-                    .filter { it.state.contractState !is MyChainableState }
-                    .map { it.ref }
-            }
-            "VALID" -> {
-                stateAndRefs.map { it.ref }
-            }
-            else -> throw IllegalArgumentException("Invalid rule type passed in")
-        }
-        val transaction = utxoLedgerService.createTransactionBuilder()
-            .setNotary(notaryLookup.notaryServices.first().name)
-            .addInputStates(inputs)
-            .addSignatories(key)
-            .addCommand(MyChainableContract.Delete())
-            .setTimeWindowUntil(Instant.now().plus(10, ChronoUnit.DAYS))
-            .toSignedTransaction()
-
-        return utxoLedgerService.finalize(transaction, emptyList()).transaction.outputStateAndRefs
-    }
-}
-
 private class Request(val command: String, val rule: String)
